requirements.txt
streamlit
msal
requests
openai
python-docx
openpyxl
PyMuPDF